Fixed mak files to get rid of warnings

+++ application_service/send_request.cc
@@ -45,7 +45,7 @@ int main(int an, char**av) {
   }
   // write request
   if (sized_socket_write(sock, serialized_request.size(), (byte*)serialized_request.data()) <
-      serialized_request.size()) {
+      (int)serialized_request.size()) {
     printf("send_request, socket write failed\n");
     return 1;
   }
